tele_bot/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==26.1.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
choicelib==0.1.5
click==8.1.7
colorama==0.4.6
Flask==3.0.3
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.19.0
multidict==6.7.1
packaging==24.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
requests==2.32.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.2.3
vbml==1.1.post1
vkbottle==4.7.0
vkbottle-types==5.199.99.12
Werkzeug==3.0.4
yarl==1.23.0
